requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via -r requirements.in
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.2.0
    # via aiohttp
cachetools==5.3.3
    # via google-auth
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
coverage[toml]==7.5.4
    # via pytest-cov
dataclasses-json==0.6.7
    # via
    #   langchain
    #   langchain-community
dnspython==2.6.1
    # via email-validator
docstring-parser==0.16
    # via google-cloud-aiplatform
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.1
    # via
    #   anyio
    #   pytest
execnet==2.1.1
    # via pytest-cache
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
flake8==7.1.0
    # via pytest-flake8
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.19.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.30.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-cloud-aiplatform==1.57.0
    # via -r requirements.in
google-cloud-bigquery==3.25.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-discoveryengine==0.11.13
    # via -r requirements.in
google-cloud-resource-manager==1.12.3
    # via google-cloud-aiplatform
google-cloud-storage==2.17.0
    # via google-cloud-aiplatform
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.63.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-resource-manager
grpcio==1.64.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   fastapi
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.1.15
    # via -r requirements.in
langchain-community==0.0.38
    # via langchain
langchain-core==0.1.52
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.0.2
    # via langchain
langsmith==0.1.82
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.3
    # via dataclasses-json
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mock==5.1.0
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   shapely
orjson==3.10.5
    # via
    #   fastapi
    #   langsmith
packaging==23.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   langchain-core
    #   marshmallow
    #   pytest
pep8==1.7.1
    # via pytest-pep8
pluggy==1.5.0
    # via pytest
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycodestyle==2.12.0
    # via flake8
pydantic==2.7.4
    # via
    #   -r requirements.in
    #   fastapi
    #   google-cloud-aiplatform
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.18.4
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pytest==8.2.2
    # via
    #   -r requirements.in
    #   pytest-cache
    #   pytest-cov
    #   pytest-flake8
    #   pytest-pep8
pytest-cache==1.0
    # via pytest-pep8
pytest-cov==5.0.0
    # via -r requirements.in
pytest-flake8==1.1.1
    # via -r requirements.in
pytest-pep8==1.0.6
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   responses
    #   uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   langchain
    #   langchain-community
    #   langsmith
    #   responses
responses==0.25.3
    # via -r requirements.in
rich==13.7.1
    # via typer
rsa==4.9
    # via google-auth
shapely==2.0.4
    # via google-cloud-aiplatform
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.31
    # via
    #   langchain
    #   langchain-community
starlette==0.37.2
    # via fastapi
tenacity==8.4.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via
    #   -r requirements.in
    #   dataclasses-json
ujson==5.10.0
    # via fastapi
urllib3==2.2.2
    # via
    #   requests
    #   responses
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
yarl==1.9.4
    # via aiohttp
